Refactor stale queued work and runtime handlers
